Rename analyzer.rs -> analyser.rs (conform spec spelling)
