+++ src/node/mod.rs
@@ -31,8 +31,8 @@ mod constant_source;
 pub use constant_source::*;
 mod panner;
 pub use panner::*;
-mod analyzer;
-pub use analyzer::*;
+mod analyser;
+pub use analyser::*;
 mod audio_buffer_source;
 pub use audio_buffer_source::*;
 mod media_element;
